Added GET req and html

+++ stock-service/src/main/java/io/pivotal/stockservice/QuotePresenter.java
@@ -2,16 +2,14 @@ package io.pivotal.stockservice;
 
 import org.springframework.stereotype.Component;
 
-import java.time.format.DateTimeFormatter;
 
 @Component
 public class QuotePresenter {
 
-    //DateTimeFormatter formatter = DateTimeFormatter.ISO_DATE_TIME;
-
+    //quote presenter is used to convery QuoteRecord objects to their string equivalent
     public QuoteInfo present(QuoteRecord record){
         return new QuoteInfo(
-                Integer.toString(record.id),
+                Long.toString(record.id),
                 record.symbol,
                 Double.toString(record.price),
                 Integer.toString(record.volume),
